requirements.txt
streamlit==1.24.0
streamlit-drawable-canvas==0.9.3
altair==4.2.2
pandas
pillow
numpy